Bump version to 1.1.5, and add a NEWS entry for this release.
Miscellaneous cleanups, mainly in the status.c code; fix a bug that
could cause heap corruption (rhncpy always clobbered the whole buffer,
but only the needed space was reserved in add_rr). This should solve
the crashes some people were seeing (this bug is not an exploitable
security hole as far as I know; the respective buffer is on the heap,
as mentioned).